mkws 18.08.26 bugfix

+++ src/controllers/main_controller.py
@@ -3,7 +3,7 @@ from PySide6.QtWidgets import *
 from PySide6.QtGui import *
 from PySide6.QtCore import *
 
-from src.views import CreateNewPassword
+from src.views.Dialogs.create_new_password import CreateNewPassword
 from src.views import Settings
 from src.views import Edit_Password
 from src.views import icons_set_color
@@ -18,6 +18,15 @@ import toml
 from datetime import datetime
 
 
+# FIXME когда удаляешь последний пароль, меню с ним должно автоматически закрываться, а не оставаться висеть, до тех пор пока не переключить на другой пароль. Из-за это не скрываются кнопки взаимодействия с паролем. А бд удалить или изменить нельзя через основное окно программы. Нужно после удаления пароля возвращать пользователя на стартовое меню. 
+# FIXME Изменить логику редактирования пароля. Нужно сделать редактирование в реальном времени, также как и создание нового пароля.
+# FIXME если пароль изменить, а после пару раз скрыть и открыть пароль, то пароль в интерфейсе заменяется на пустое значение, при этом он меняется только в интерфейсе, в базе данных он не изменяется. Дело в функции скрытия пароля.
+
+# TODO Добавление стилей для диалоговых окон. 
+# TODO Создать подтверждения пароля от базы данных.  
+# TODO добавить окно с информацией о бд. Чтобы зайти в это окно, нужно будет ввести логин и пароль от базы данных. Там же их можно будет и поменять. 
+# TODO Сделать уведомления, которые будут всплывать, когда пользователь будет взаимодействовать с программой: "Пароль успешно создан", "Пароль успешно удален". Но эти уведомления должны быть всплывающими в интерфейсе самой программы, а не системно. Перекрывая часть интерфейса(пример vscode). 
+# TODO Сделать поиск
 
 class MainController():
     def __init__(self, view):
@@ -27,7 +36,7 @@ class MainController():
             logger.success("Config successfully loaded ✅")
             self.config = toml.load(config_file)
 
-        self._view  = view      #MainWindow
+        self._view = view #MainWindow
 
         self.service_name_ = None
         self.login_ = None
@@ -35,28 +44,32 @@ class MainController():
         self.password_ = None
         self.current_password = None
 
-        self._connect_signals() #Signals
-        self.HotKeys()
+        self._connect_signals() #Connects
+        self.HotKeys() # HotKeys
 
 
+    # connect signals
     def _connect_signals(self):
         
         self._view.tree_view.clicked.connect(self._on_category_clicked)               #tree select category
-        self._view.new_password_btn.clicked.connect(self._new_password_clicked)       #new password button
-        self._view.settings_btn.clicked.connect(self._setting_clicked)                #settings button
-        self._view.search_btn.clicked.connect(self._search_clicked)
-        self._view.edit_password_btn.clicked.connect(self._edit_password_clicked)     #edit password button 
-        self._view.hide_password_btn.clicked.connect(self.hide)
-        self._view.del_password_btn.clicked.connect(self._del_password_clicked)       #del password button 
+        self._view.new_password_btn.clicked.connect(self._new_password_clicked)       #new password menu(button)
+        self._view.settings_btn.clicked.connect(self._setting_clicked)                #settings menu(button)
+        self._view.edit_password_btn.clicked.connect(self._edit_password_clicked)     #edit password menu(button)
+        self._view.hide_password_btn.clicked.connect(self.hide)                       #hide password(button)
+        self._view.del_password_btn.clicked.connect(self._del_password_clicked)       #del password(button)
+
+        self._view.new_password_menu.triggered.connect(self._new_password_clicked)    #new password menu(menubar)
+        self._view.edit_password_menu.triggered.connect(self._edit_password_clicked)
+
 
         # TODO Нужно блокировать окно программы, когда открыто диалоговое окно, чтобы нельзя было открыть одновремено несколько окон добавления пароля или окон редактрирования. 
 
-    def _on_category_clicked(self, index): #select category
+    # select category
+    def _on_category_clicked(self, index): 
 
-
-        self.item = self._view.tree_model.itemFromIndex(index)
-        item_name = self.item.text()                                          #get text
-        self._view.title.setText(f"{item_name}".upper())                      #change title
+        self.item = self._view.tree_model.itemFromIndex(index) #current element
+        item_name = self.item.text() #get text
+        self._view.title.setText(f"{item_name}".upper()) #change title
         
 
         if self.item == self._view.root_item: # если выбран заголовок в пункте меню, то скрывать значения сервиса логина и пароля, вызвав функцию для просмотра информации о бд
@@ -72,6 +85,8 @@ class MainController():
             self._view.edit_password_btn.hide()
             self._view.del_password_btn.hide()
             self._view.welcome_logo.hide()
+
+            self._view.edit_password_menu.blockSignals(True)
 
             self.db_information()
             return
@@ -100,6 +115,11 @@ class MainController():
             self._view.db_login.hide()
             self._view.db_password.hide()
 
+            #unblock signals
+            self._view.edit_password_menu.blockSignals(False)
+            self._view.del_password_btn.setEnabled(True)
+
+
             if self.config["privacy"]["auto_hide_passwords"]:
                 self._view.hide_password_btn_state = False
             
@@ -125,21 +145,22 @@ class MainController():
         if self.config["privacy"]["auto_hide_passwords"]:
             self.hide()
         else:
-            ...
-            
+            return
 
-
-    def _new_password_clicked(self): #add new password
+    # add new password
+    def _new_password_clicked(self):
         logger.debug("New button clicked")
 
-        self.new_menu = CreateNewPassword(self._view)
+        self.new_window = CreateNewPassword(self._view)
 
         self._new_password_saved = False 
         self._new_tree_item = None
         self._current_new_id = None
 
-        self.new_menu.data_created_password.connect(self._update_ui_new_password_realtime)
+        self.new_window.data_created_password.connect(self._update_ui_new_password_realtime)
+        self.new_window.exec()
 
+    # add new password realtime update ui
     def _update_ui_new_password_realtime(self, data: dict):
         get_service = data.get("service", "")
         get_url = data.get("url", "")
@@ -158,16 +179,15 @@ class MainController():
                 self._new_password_saved = True
                 self._current_new_id = new_id
                 
-                current_date = datetime.now().strftime("%Y-%m-%d %H:%M")
-
-                # Создаем элемент в дереве
+                current_date = datetime.now().strftime("%Y-%m-%d %H:%M") # get current date
+                # Create element on tree
                 service_item = QStandardItem(get_service or "New Entry")
-                service_item.setData(self._current_new_id, Qt.ItemDataRole.UserRole)
-                service_item.setData(get_login, Qt.ItemDataRole.UserRole + 1)
-                service_item.setData(get_url, Qt.ItemDataRole.UserRole + 2)
-                service_item.setData(get_password, Qt.ItemDataRole.UserRole + 3)
-                service_item.setData(current_date, Qt.ItemDataRole.UserRole + 4) # Ставим текущую дату
-                service_item.setData("", Qt.ItemDataRole.UserRole + 5)           # Пустое описание
+                service_item.setData(self._current_new_id, Qt.ItemDataRole.UserRole) # current id
+                service_item.setData(get_login, Qt.ItemDataRole.UserRole + 1)    # current login
+                service_item.setData(get_url, Qt.ItemDataRole.UserRole + 2)      # current url 
+                service_item.setData(get_password, Qt.ItemDataRole.UserRole + 3) # current password
+                service_item.setData(current_date, Qt.ItemDataRole.UserRole + 4) # current date 
+                service_item.setData("", Qt.ItemDataRole.UserRole + 5)           
 
                 self._view.root_item.appendRow(service_item)
                 self._new_tree_item = service_item
@@ -187,42 +207,44 @@ class MainController():
                 self._new_tree_item.setData(get_login, Qt.ItemDataRole.UserRole + 1)
                 self._new_tree_item.setData(get_url, Qt.ItemDataRole.UserRole + 2)
                 self._new_tree_item.setData(get_password, Qt.ItemDataRole.UserRole + 3)
-                logger.debug(f"Password ID {self._current_new_id} updated in DB and UI") 
- 
-    def _edit_password_clicked(self): #edit password
-        #open the child window to edit a password
+
+    # edit password
+    def _edit_password_clicked(self):
         logger.debug("Edit button clicked")
         
-        self.edit_menu = Edit_Password()
-        # self.edit_menu.exec()
+        self.edit_window = Edit_Password()
         
         get_service = self.service_name_
         get_url = self.url_
         get_login = self.login_
         get_password = self.password_
 
-        # заполнять холдертекст старыми данными
-        self.edit_menu.edit_input_service.setPlaceholderText(get_service)
-        self.edit_menu.edit_input_login.setPlaceholderText(get_login)
-        self.edit_menu.edit_input_url.setPlaceholderText(get_url)
-        self.edit_menu.edit_input_password.setPlaceholderText(get_password)
+        # fild holdertext old data
+        self.edit_window.edit_input_service.setPlaceholderText(get_service)
+        self.edit_window.edit_input_login.setPlaceholderText(get_login)
+        self.edit_window.edit_input_url.setPlaceholderText(get_url)
+        self.edit_window.edit_input_password.setPlaceholderText(get_password)
 
-        self.edit_menu.data_edit_password.connect(self.update_ui_edited)
-    
-    def update_ui_edited(self, data:dict): #update edited passwords
+        # get signal
+        self.edit_window.data_edit_password.connect(self._update_ui_edited)
+
+        self.edit_window.exec() # execute edit window
+
+    # update edited passwords
+    def _update_ui_edited(self, data:dict):
         #get edited password 
         new_service  = data.get("service") or self.service_name_
         new_url      = data.get("url") or self.url_
         new_login    = data.get("login") or self.login_
         new_password = data.get("password") or self.password_
         
-        if new_service != "": logger.debug(f"Received updated service data: {new_service}")
-        if new_url != "": logger.debug(f"Received updated url data: {new_url}")
-        if new_login != "": logger.debug(f"Received updated login data: {new_login}")
-        if new_password != "": logger.debug(f"Received updated password data: {new_password}")
+        if new_service != "": logger.debug(f"Received updated service data: {new_service}") # if service = Null return old data
+        if new_url != "": logger.debug(f"Received updated url data: {new_url}") # if url = Null return old data
+        if new_login != "": logger.debug(f"Received updated login data: {new_login}") # if login = Null return old data
+        if new_password != "": logger.debug(f"Received updated password data: {new_password}")# if password = Null return old data
 
 
-        #change in db
+        # change in db
         success = database.update_password(
             password_id=self._view.entry_id,
             service=new_service,
@@ -242,46 +264,67 @@ class MainController():
             self.login_ = new_login
             self.password_ = new_password
 
-            self.item.setText(self.service_name_)                     # заголовок в дереве
-            self._view.title.setText(f"{self.service_name_}".upper()) #заголовок в right_workspace
-            
-    def _del_password_clicked(self, tree_view): #del password
+            self.item.setText(self.service_name_)                     # title on tree 
+            self._view.title.setText(f"{self.service_name_}".upper()) # title on right_workspace
+
+    # delete password
+    def _del_password_clicked(self, tree_view): 
         logger.debug("Del button clicked")
-        # Получаем индекс выбранного элемента
-        current_index = self._view.tree_view.currentIndex()
+
+        reply = QMessageBox.question(
+            None, 
+            "Confirmation", 
+            "Are you sure you want to delete the password?",
+            QMessageBox.Yes | QMessageBox.No
+        )
         
-        if not current_index.isValid():
-            return
-
-        model = self._view.tree_view.model() 
-        row = current_index.row()
-        parent_index = current_index.parent()
-
-        # Находим индекс первой колонки (0) для выделенной строки
-        first_column_index = model.index(row, 0, parent_index)
-        
-        # Извлекаем спрятанный ID базы данных из роли UserRole
-        db_id = model.data(first_column_index, Qt.ItemDataRole.UserRole)
-
-        if db_id is None:
-            logger.error("The database identifier is missing from the interface element!")
-            return
-
-        # Удаления из базы данных
-        if database.delete_password(db_id):
-            # Если из файла базы данных удалено, удаляем строку из интерфейса
-            model.removeRow(row, parent_index)
-            logger.success(f"The row with ID {db_id} has been removed from the interface.")
+        if reply == QMessageBox.Yes:
+            logger.debug("Password deleted.")
+            #Get current index element
+            current_index = self._view.tree_view.currentIndex()
     
-    def _setting_clicked(self): #settings
+            if not current_index.isValid():
+                return
+    
+            model = self._view.tree_view.model() 
+            row = current_index.row()
+            parent_index = current_index.parent()
+    
+            # index first row
+            first_column_index = model.index(row, 0, parent_index)
+            
+            #get database id 
+            db_id = model.data(first_column_index, Qt.ItemDataRole.UserRole)
+    
+            if db_id is None:
+                logger.error("The database identifier is missing from the interface element!")
+                return
+    
+            #delete
+            if database.delete_password(db_id):
+                model.removeRow(row, parent_index)
+                logger.success(f"The row with ID {db_id} has been removed from the interface.")
+        else:
+            logger.debug("Deletion cancelled.")
+            return      
+
+    # settings
+    def _setting_clicked(self): 
         #open the child window for application settings
         logger.debug("Settings button clicked")
         Settings(self._view)
 
-    def _search_clicked(self): #seach
-        logger.debug("Search button clicked")
+    # seach
+    def _search(self, data):
+        # logger.debug("Search button clicked")
+        search_input_field = data.get("text", "")
 
-    def db_information(self): #database information
+        print(search_input_field)
+
+        #TODO нужно разделить искомое слово на отдельные символы и сравнивать их с service, url, login или #tag. Скрывать неподходящие под поиск пароли и оставлять в дереве, только то, что подходит. 
+
+    # database information
+    def db_information(self):
         logger.debug("db_infromation clicked")
         self._view.db_size_lb.show()
         self._view.db_creation_date.show()
@@ -290,9 +333,9 @@ class MainController():
         self._view.db_count_dublicate.show()
         self._view.db_login.show()
         self._view.db_password.show()
-    
-    def hide(self, checked=None): #hide passwords
-        #FIXME сделать так, чтобы когда кнопка нажималась один раз скрытие не спадала и на других паролях тоже
+
+    # hide/show passwords
+    def hide(self, checked=None): 
         icon_size = QSize(24, 24)
     
         if not self._view.hide_password_btn_state:
@@ -313,13 +356,15 @@ class MainController():
             self._view.hide_password_btn_state = False
             logger.debug("Password is shown")
 
-    def HotKeys(self): #Hotkeys
+    # hotkeys
+    def HotKeys(self): 
         hotkeys = {
             "new_password":  self.config["hotkeys"].get("new_password", "Ctrl+N"),
             "edit_password": self.config["hotkeys"].get("edit_password", "Ctrl+E"),
             "del_password": self.config["hotkeys"].get("del_password", "Ctrl+Backspace"),
             "hide_password": self.config["hotkeys"].get("hide_password", "Ctrl+G"),
-            "settings": self.config["hotkeys"].get("settings", "Ctrl+I")
+            "settings": self.config["hotkeys"].get("settings", "Ctrl+I"),
+            "main_menu": self.config["hotkeys"].get("main_menu", "Escape")
         }
 
         if hotkeys["new_password"]:
@@ -342,3 +387,23 @@ class MainController():
             self.shortcut_settings = QShortcut(QKeySequence(hotkeys["settings"]), self._view)
             self.shortcut_settings.activated.connect(self._setting_clicked)
 
+        if hotkeys["main_menu"]:
+            self.shortcut_main_menu = QShortcut(QKeySequence(hotkeys["main_menu"]), self._view)
+            self.shortcut_main_menu.activated.connect(self.main_menu)
+
+    def main_menu(self):
+
+        # hide all elements right workspace
+        self._view.title.hide()
+        self._view.service_lb.hide()
+        self._view.url_lb.hide()
+        self._view.login_lb.hide()
+        self._view.password_lb.hide()
+        self._view.creation_date_lb.hide()
+        self._view.description_lb.hide()
+        self._view.description.hide()
+        self._view.hide_password_btn.hide()
+        self._view.edit_password_btn.hide()
+        self._view.del_password_btn.hide()
+        # show logo
+        self._view.welcome_logo.show()
